requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.9
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
